# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MarketPulse API v2",
    "description": "EOD swing-trading data for NSE stocks. Envelope {as_of, freshness, total, returned, rows, meta}; NULL stays NULL; every read accepts as_of (time travel).",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/v2/market/regime": {
      "get": {
        "summary": "Market Regime",
        "operationId": "market_regime_api_v2_market_regime_get",
        "parameters": [
          {
            "name": "as_of",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Time travel: all data bounded to sessions on or before this date (YYYY-MM-DD)",
              "title": "As Of"
            },
            "description": "Time travel: all data bounded to sessions on or before this date (YYYY-MM-DD)"
          },
          {
            "name": "days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 2000,
              "minimum": 1,
              "default": 126,
              "title": "Days"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 5000,
              "minimum": 1,
              "default": 500,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_RegimeRow_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v2/desk/queues": {
      "get": {
        "summary": "Desk Queues",
        "operationId": "desk_queues_api_v2_desk_queues_get",
        "parameters": [
          {
            "name": "as_of",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Time travel: all data bounded to sessions on or before this date (YYYY-MM-DD)",
              "title": "As Of"
            },
            "description": "Time travel: all data bounded to sessions on or before this date (YYYY-MM-DD)"
          },
          {
            "name": "all_timeframes",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Also count W/M (slow until setup_daily exists)",
              "default": false,
              "title": "All Timeframes"
            },
            "description": "Also count W/M (slow until setup_daily exists)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_QueueSummaryRow_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/desk/queue/{name}": {
      "get": {
        "summary": "Desk Queue",
        "operationId": "desk_queue_api_v2_desk_queue__name__get",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "enum": [
                "darvas_squeeze",
                "darvas_10ema",
                "vcp"
              ],
              "type": "string",
              "title": "Name"
            }
          },
          {
            "name": "as_of",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Time travel: all data bounded to sessions on or before this date (YYYY-MM-DD)",
              "title": "As Of"
            },
            "description": "Time travel: all data bounded to sessions on or before this date (YYYY-MM-DD)"
          },
          {
            "name": "tf",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "D",
                "W",
                "M"
              ],
              "type": "string",
              "default": "D",
              "title": "Tf"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "rs_percentile",
                    "distance_to_trigger_pct",
                    "risk_pct",
                    "setup_age_sessions",
                    "rvol",
                    "change_1d_pct",
                    "squeeze_pct",
                    "excess_vs_midsml400_63d",
                    "symbol"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sort"
            }
          },
          {
            "name": "desc",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Desc"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 5000,
              "minimum": 1,
              "default": 500,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_QueueRow_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/desk/watchlist": {
      "get": {
        "summary": "Desk Watchlist",
        "operationId": "desk_watchlist_api_v2_desk_watchlist_get",
        "parameters": [
          {
            "name": "as_of",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Time travel: all data bounded to sessions on or before this date (YYYY-MM-DD)",
              "title": "As Of"
            },
            "description": "Time travel: all data bounded to sessions on or before this date (YYYY-MM-DD)"
          },
          {
            "name": "symbols",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 25000,
              "description": "Comma-separated symbols, in display order",
              "default": "",
              "title": "Symbols"
            },
            "description": "Comma-separated symbols, in display order"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 5000,
              "minimum": 1,
              "default": 500,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_DeskWatchRow_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v2/research/analogs": {
      "get": {
        "summary": "Research Analogs",
        "operationId": "research_analogs_api_v2_research_analogs_get",
        "parameters": [
          {
            "name": "as_of",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Time travel: all data bounded to sessions on or before this date (YYYY-MM-DD)",
              "title": "As Of"
            },
            "description": "Time travel: all data bounded to sessions on or before this date (YYYY-MM-DD)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_MarketAnalogRow_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Envelope_RegimeRow_": {
        "properties": {
          "as_of": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "As Of"
          },
          "freshness": {
            "$ref": "#/components/schemas/Freshness"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "returned": {
            "type": "integer",
            "title": "Returned"
          },
          "rows": {
            "items": {
              "$ref": "#/components/schemas/RegimeRow"
            },
            "type": "array",
            "title": "Rows"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "type": "object",
        "required": [
          "freshness",
          "total",
          "returned",
          "rows",
          "meta"
        ],
        "title": "Envelope[RegimeRow]"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    }
  }
}